pyproject.toml
[project]
name = "oncall-agents"                                                                            
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",                                                                            
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",                                                                 
    "langgraph>=0.2.0",
    "langchain-milvus>=0.1.0",                                                                    
    "langchain-voyageai>=0.1.0",
    "pymilvus>=2.4.0",                                                                            
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",                                                                       
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",                                                                              
    "httpx>=0.27.0",                                                                              
]

[dependency-groups]                                                                                          
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]                                                                                    
testpaths = ["tests"]                                                                                        
markers = [                                 
    "integration: marks tests that hit real APIs (slow)",
]